common titlepage data
