+++ modules/ui/src/components/TitlePage/TitlePage.css
@@ -4,13 +4,6 @@
 }
 
 .titlePageWelcomeContainer {
-  position: absolute;
-  z-index: 100;
-  top: 30px;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: 15px;
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
