+++ modules/ui/src/components/TitlePage/TitlePage.tsx
@@ -8,15 +8,17 @@ import { Page } from "../Page";
 import { TitleBoard } from "./TitlePage.Board";
 import { TitlePageBody } from "./TitlePage.Body";
 import { setTitlePageAttributes } from "./setTitlePageAttributes";
+import { useTitleData } from "./TitlePage.useTitleData";
 
 export const TitlePage: AlgolPage = props => {
   const { actions } = props;
+  const titleDemo = useTitleData();
   return (
     <Fragment>
       <Page
         title="Hello!"
-        top={<TitleBoard actions={actions} />}
-        body={<TitlePageBody actions={actions} />}
+        top={<TitleBoard titleData={titleDemo.titleData} />}
+        body={<TitlePageBody actions={actions} titleDemo={titleDemo} />}
       />
     </Fragment>
   );
